finished

+++ apps/books/models.py
@@ -2,6 +2,7 @@ from __future__ import unicode_literals
 from django.db import models
 from ..users.models import *
 # Create your models here.
+#------------------------------------------------------------------------
 class ReviewManager(models.Manager):
     def basic_validator(self, postData):
         errors = {}
@@ -9,11 +10,12 @@ class ReviewManager(models.Manager):
             errors["content"] = "content cannot be less than 10 characters"
         if len(postData['content']) > 200:
             errors["content"] = "content cannot be over 200 characters"
-        if int(postData['rating'])>5 :
+        if float(postData['rating'])>5 :
             errors["rating"] = "you cannot give a higher rating than 5"
-        if int(postData['rating'])%.5 != 0:
+        if float(postData['rating'])%.5 != 0:
             errors["rating"] = "you can only give ratings divisible by 0.5"
         return errors
+#------------------------------------------------------------------------
 class BookManager(models.Manager):
     def basic_validator(self, postData):
         errors = {}
@@ -30,6 +32,7 @@ class BookManager(models.Manager):
         if len(postData['desc']) > 200:
             errors["desc"] = "description cannot be over 200 characters"
         return errors
+#------------------------------------------------------------------------    
 class Book(models.Model):
     title = models.CharField(max_length=255)
     author = models.CharField(max_length=255)
@@ -37,11 +40,13 @@ class Book(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     objects=BookManager()
+#------------------------------------------------------------------------
 class Review(models.Model):
     user=models.ForeignKey(User, related_name="reviews")
     book=models.ForeignKey(Book, related_name='books')
-    rating=models.IntegerField()
+    rating=models.FloatField()
     content = models.CharField(max_length=255)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
     objects=ReviewManager()
+#------------------------------------------------------------------------
